05-data-platforms/zoomcamp/pipeline/assets/ingestion/requirements.txt
# Python dependencies required by the python ingestion asset.
# Keep minimal and pin versions for reproducibility.
# Docs: https://getbruin.com/docs/bruin/assets/python

pandas==2.2.3
requests==2.31.0
pyarrow==12.0.1
python-dateutil==2.8.2
numpy<2

